requirements_dev.txt
-r requirements.txt
isort==5.13.2
django-stubs==4.2.7
django-stubs-ext==4.2.7
flake8==7.0.0
flake8-print==5.0.0
mypy==1.9.0
mypy-extensions==1.0.0
